+++ onnx/main.py
@@ -3,6 +3,7 @@ import os
 os.environ['KIVY_GL_BACKEND'] = 'sdl2'
 import sys
 from threading import Thread
+import requests
 
 from kivy.lang import Builder
 from kivy.properties import StringProperty, NumericProperty, ObjectProperty
@@ -11,6 +12,7 @@ from kivy.metrics import dp, sp
 from kivy.utils import platform
 from kivy.uix.image import Image
 from kivy.uix.camera import Camera
+from kivy.clock import Clock
 
 from kivymd.app import MDApp
 from kivymd.uix.navigationdrawer import MDNavigationDrawerMenu
@@ -26,10 +28,14 @@ Window.softinput_mode = "below_target"
 from screens.img_obj_detect import ImgObjDetBox, TempSpinWait
 from screens.cam_obj_detect import CamObjDetBox
 from screens.setting import SettingsBox
-from onnx_vision import OnnxDetect
+from onnx_detect import OnnxDetect
+from onnx_classify import OnnxClassify
 
 ## Global definitions
-__version__ = "0.0.4"
+__version__ = "0.0.5" # The APP version
+
+detect_model_url = "https://github.com/onnx/models/raw/main/validated/vision/object_detection_segmentation/ssd-mobilenetv1/model/ssd_mobilenet_v1_10.onnx"
+classify_model_url = "https://github.com/onnx/models/raw/main/validated/vision/classification/resnet/model/resnet18-v1-7.onnx"
 # Determine the base path for your application's resources
 if getattr(sys, 'frozen', False):
     # Running as a PyInstaller bundle
@@ -47,12 +53,18 @@ class ContentNavigationDrawer(MDNavigationDrawerMenu):
 
 ## kivymd app class
 class VisionAiApp(MDApp):
-    is_onnx_running = ObjectProperty()
+    is_detect_running = ObjectProperty(None)
+    is_downloading = ObjectProperty(None)
+    is_classify_running = ObjectProperty(None)
     image_path = StringProperty("")
     op_img_path = StringProperty("")
     onnx_detect = ObjectProperty(None)
+    onnx_detect_sess = ObjectProperty(None)
+    onnx_classify = ObjectProperty(None)
+    onnx_classify_sess = ObjectProperty(None)
     cam_found = ObjectProperty(None)
     camera = ObjectProperty(None)
+    detect_model_path = StringProperty("")
 
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
@@ -103,12 +115,8 @@ class VisionAiApp(MDApp):
             self.op_dir = os.path.join(self.user_data_dir, 'outputs')
         os.makedirs(self.model_dir, exist_ok=True)
         os.makedirs(self.op_dir, exist_ok=True)
-
-        # create onnx objects
-        self.onnx_detect = OnnxDetect(
-            save_dir=self.op_dir,
-            model_dir=self.model_dir,
-        )
+        self.detect_model_path = os.path.join(self.model_dir, "ssd_mobilenet_v1_10.onnx")
+        self.classify_model_path = os.path.join(self.model_dir, "resnet18-v1-7.onnx")
 
         # file managers
         self.is_img_manager_open = False
@@ -126,7 +134,114 @@ class VisionAiApp(MDApp):
             select_path=self.select_op_path,
             selector="folder",  # Restrict to selecting directories only
         )
+
+        if not os.path.exists(self.detect_model_path):
+            self.popup_detect_model()
+
+        # create onnx objects
+        self.onnx_detect = OnnxDetect(
+            save_dir=self.op_dir,
+            model_dir=self.model_dir,
+        )
+
         print("Initialisation is successfull")
+
+    def update_download_progress(self, downloaded, total_size):
+        if total_size > 0:
+            percentage = (downloaded / total_size) * 100
+            self.download_progress.text = f"Progress: {percentage:.1f}%"
+        else:
+            self.download_progress.text = f"Progress: {downloaded} bytes"
+
+    def download_file(self, download_url, download_path):
+        filename = download_url.split("/")[-1]
+        try:
+            self.is_downloading = filename
+            with requests.get(download_url, stream=True) as req:
+                req.raise_for_status()
+                total_size = int(req.headers.get('content-length', 0))
+                downloaded = 0
+                with open(download_path, 'wb') as f:
+                    for chunk in req.iter_content(chunk_size=8192):
+                        if chunk:
+                            f.write(chunk)
+                            downloaded += len(chunk)
+                            Clock.schedule_once(lambda dt: self.update_download_progress(downloaded, total_size))
+            if os.path.exists(download_path):
+                Clock.schedule_once(lambda dt: self.show_toast_msg(f"Download complete: {download_path}"))
+            else:
+                Clock.schedule_once(lambda dt: self.show_toast_msg(f"Download failed for: {download_path}", is_error=True))
+            self.is_downloading = False
+        except requests.exceptions.RequestException as e:
+            print(f"Error downloading the onnx file: {e} 😞")
+            Clock.schedule_once(lambda dt: self.show_toast_msg(f"Download failed for: {download_path}", is_error=True))
+            self.is_downloading = False
+
+    def download_model_file(self, model_url, download_path, instance=None):
+        self.txt_dialog_closer(instance)
+        filename = download_path.split("/")[-1]
+        print(f"Starting the download for: {filename}")
+        if self.root.ids.screen_manager.current == "imgObjDetect":
+            result_box = self.root.ids.img_detect_box.ids.result_image
+        elif self.root.ids.screen_manager.current == "imgClassify":
+            result_box = self.root.ids.img_classify_box.ids.result_label
+        else:
+            result_box = self.root.ids.cam_detect_box.ids.cam_result_image
+        result_box.clear_widgets()
+        self.download_progress = MDLabel(
+            text="Progress: 0%",
+            halign="center"
+        )
+        result_box.add_widget(self.download_progress)
+        Thread(target=self.download_file, args=(model_url, download_path), daemon=True).start()
+
+    def download_detect_model(self, instance):
+        self.download_model_file(detect_model_url, self.detect_model_path, instance)
+
+    def download_classify_model(self, instance):
+        self.download_model_file(classify_model_url, self.classify_model_path, instance)
+
+    def popup_detect_model(self):
+        buttons = [
+            MDFlatButton(
+                text="Cancel",
+                theme_text_color="Custom",
+                text_color=self.theme_cls.primary_color,
+                on_release=self.txt_dialog_closer
+            ),
+            MDFlatButton(
+                text="Ok",
+                theme_text_color="Custom",
+                text_color="green",
+                on_release=self.download_detect_model
+            ),
+        ]
+        self.show_text_dialog(
+            "Downlaod the model file",
+            f"You need to downlaod the file for the first time (~30MB)",
+            buttons
+        )
+
+    def popup_classify_model(self):
+        buttons = [
+            MDFlatButton(
+                text="Cancel",
+                theme_text_color="Custom",
+                text_color=self.theme_cls.primary_color,
+                on_release=self.txt_dialog_closer
+            ),
+            MDFlatButton(
+                text="Ok",
+                theme_text_color="Custom",
+                text_color="green",
+                on_release=self.download_classify_model
+            ),
+        ]
+        self.show_text_dialog(
+            "Downlaod the model file",
+            f"You need to downlaod the file for the first time (~45MB)",
+            buttons
+        )
 
     def show_toast_msg(self, message, is_error=False):
         from kivymd.uix.snackbar import MDSnackbar
@@ -158,6 +273,8 @@ class VisionAiApp(MDApp):
         self.show_toast_msg("You can detect objects on your image!")
 
     def on_cam_obj_detect(self):
+        if not os.path.exists(self.detect_model_path) and self.is_downloading != "ssd_mobilenet_v1_10.onnx":
+            self.popup_detect_model()
         self.show_toast_msg("Capture an image & detect objects!")
         self.cam_uix = self.root.ids.cam_detect_box.ids.capture_image
         self.cam_uix.clear_widgets()
@@ -192,6 +309,16 @@ class VisionAiApp(MDApp):
             print(f"Error setting up the camera: {e}")
             self.show_toast_msg(f"Error setting up the camera: {e}", is_error=True)
 
+    def on_img_classify(self):
+        if not os.path.exists(self.classify_model_path) and self.is_downloading != "resnet18-v1-7.onnx":
+            self.popup_classify_model()
+        self.show_toast_msg("Select and image & get top 5 predictions")
+        if not self.onnx_classify:
+            self.onnx_classify = OnnxClassify(
+                save_dir=self.op_dir,
+                model_dir=self.model_dir,
+            )
+
     def on_cam_obj_dt_leave(self):
         if self.cam_found:
             self.camera.play = False
@@ -199,25 +326,60 @@ class VisionAiApp(MDApp):
 
     def open_img_file_manager(self):
         """Open the file manager to select an image file. On android use Downloads or Pictures folders only"""
-        if self.is_onnx_running:
+        if self.is_downloading == "ssd_mobilenet_v1_10.onnx":
+            self.show_toast_msg("Please wait for the model download to finish!", is_error=True)
+            return
+        if not os.path.exists(self.detect_model_path) and self.is_downloading != "ssd_mobilenet_v1_10.onnx":
+            self.onnx_detect_sess = False
+            self.popup_detect_model()
+            return
+        if not self.onnx_detect_sess:
+            self.onnx_detect_sess = self.onnx_detect.start_detect_session()
+        if self.is_detect_running:
             self.show_toast_msg("Please wait for the current operation to finish", is_error=True)
             return
         try:
-            self.img_file_manager.show(self.internal_storage)  # native app specific path
+            self.img_file_manager.show(self.external_storage)  # native app specific path
+            self.is_img_manager_open = True
+        except Exception as e:
+            self.show_toast_msg(f"Error: {e}", is_error=True)
+
+    def open_clsfy_img_file(self):
+        """Open the file manager to select an image file. On android use Downloads or Pictures folders only"""
+        if self.is_downloading == "resnet18-v1-7.onnx":
+            self.show_toast_msg("Please wait for the model download to finish!", is_error=True)
+            return
+        if not os.path.exists(self.classify_model_path) and self.is_downloading != "resnet18-v1-7.onnx":
+            self.onnx_classify_sess = False
+            self.popup_classify_model()
+            return
+        if not self.onnx_classify_sess: # update it
+            self.onnx_classify_sess = self.onnx_classify.start_classify_session()
+        if self.is_classify_running:
+            self.show_toast_msg("Please wait for the current operation to finish", is_error=True)
+            return
+        try:
+            self.img_file_manager.show(self.external_storage)
             self.is_img_manager_open = True
         except Exception as e:
             self.show_toast_msg(f"Error: {e}", is_error=True)
 
     def select_img_path(self, path: str):
         self.image_path = path
-        uploaded_image_box = self.root.ids.img_detect_box.ids.uploaded_image
+        if self.root.ids.screen_manager.current == "imgObjDetect":
+            uploaded_image_box = self.root.ids.img_detect_box.ids.uploaded_image
+        else:
+            uploaded_image_box = self.root.ids.img_classify_box.ids.uploaded_image
         uploaded_image_box.clear_widgets()
         fitImage = Image(
             source = path,
             fit_mode = "contain"
         )
         uploaded_image_box.add_widget(fitImage)
-        result_box = self.root.ids.img_detect_box.ids.result_image
+        if self.root.ids.screen_manager.current == "imgObjDetect":
+            result_box = self.root.ids.img_detect_box.ids.result_image
+        else:
+            result_box = self.root.ids.img_classify_box.ids.result_label
         result_box.clear_widgets()
         self.img_file_exit_manager()
 
@@ -272,12 +434,15 @@ class VisionAiApp(MDApp):
         if self.image_path == "":
             self.show_toast_msg("No image is selected!", is_error=True)
             return
-        if self.is_onnx_running:
+        if not self.onnx_detect_sess:
+            self.onnx_detect_sess = self.onnx_detect.start_detect_session()
+            self.submit_onnx_detect()
+        if self.is_detect_running:
             self.show_toast_msg("Please wait for the previous request to finish", is_error=True)
             return
         onnx_thread = Thread(target=self.onnx_detect.run_detect, args=(self.image_path, self.onnx_detect_callback, "imgObjDetect"), daemon=True)
         onnx_thread.start()
-        self.is_onnx_running = True
+        self.is_detect_running = True
         tmp_spin = TempSpinWait()
         result_box = self.root.ids.img_detect_box.ids.result_image
         result_box.clear_widgets()
@@ -287,7 +452,17 @@ class VisionAiApp(MDApp):
         if not self.cam_found:
             self.show_toast_msg("Camera could not be loaded!", is_error=True)
             return
-        if self.is_onnx_running:
+        if self.is_downloading == "ssd_mobilenet_v1_10.onnx":
+            self.show_toast_msg("Please wait for the model download to finish!", is_error=True)
+            return
+        if not os.path.exists(self.detect_model_path) and self.is_downloading != "ssd_mobilenet_v1_10.onnx":
+            self.onnx_detect_sess = False
+            self.popup_detect_model()
+            return
+        if not self.onnx_detect_sess:
+            self.onnx_detect_sess = self.onnx_detect.start_detect_session()
+            self.capture_n_onnx_detect()
+        if self.is_detect_running:
             self.show_toast_msg("Please wait for the previous request to finish", is_error=True)
             return
         self.image_path = ""
@@ -301,9 +476,27 @@ class VisionAiApp(MDApp):
         self.camera.export_to_png(self.image_path)
         onnx_thread = Thread(target=self.onnx_detect.run_detect, args=(self.image_path, self.onnx_detect_callback, "camObjDetect"), daemon=True)
         onnx_thread.start()
-        self.is_onnx_running = True
+        self.is_detect_running = True
         tmp_spin = TempSpinWait()
         result_box = self.root.ids.cam_detect_box.ids.cam_result_image
+        result_box.clear_widgets()
+        result_box.add_widget(tmp_spin)
+
+    def submit_onnx_classify(self):
+        if self.image_path == "":
+            self.show_toast_msg("No image is selected!", is_error=True)
+            return
+        if not self.onnx_classify_sess:
+            self.onnx_classify_sess = self.onnx_classify.start_classify_session()
+            self.submit_onnx_classify()
+        if self.is_classify_running:
+            self.show_toast_msg("Please wait for the previous request to finish", is_error=True)
+            return
+        onnx_thread = Thread(target=self.onnx_classify.run_classify, args=(self.image_path, self.onnx_classify_callback, "imgClassify"), daemon=True)
+        onnx_thread.start()
+        self.is_classify_running = True
+        tmp_spin = TempSpinWait()
+        result_box = self.root.ids.img_classify_box.ids.result_label
         result_box.clear_widgets()
         result_box.add_widget(tmp_spin)
 
@@ -311,7 +504,7 @@ class VisionAiApp(MDApp):
         status = onnx_resp["status"]
         message = onnx_resp["message"]
         caller = onnx_resp["caller"]
-        self.is_onnx_running = False
+        self.is_detect_running = False
         if caller == "camObjDetect":
             result_box = self.root.ids.cam_detect_box.ids.cam_result_image
         else:
@@ -337,6 +530,24 @@ class VisionAiApp(MDApp):
         else:
             self.show_toast_msg(message, is_error=True)
 
+    def onnx_classify_callback(self, onnx_resp):
+        status = onnx_resp["status"]
+        message = onnx_resp["message"]
+        caller = onnx_resp["caller"]
+        self.is_classify_running = False
+        result_box = self.root.ids.img_classify_box.ids.result_label
+        result_box.clear_widgets()
+        if status is True:
+            classify_label = MDLabel(
+                text=message,
+                halign="left",
+                valign="top",
+                markup=True
+            )
+            result_box.add_widget(classify_label)
+        else:
+            self.show_toast_msg(message, is_error=True)
+
     def reset_object_detect(self):
         self.image_path = ""
         uploaded_image_box = self.root.ids.img_detect_box.ids.uploaded_image
@@ -347,6 +558,13 @@ class VisionAiApp(MDApp):
     def reset_cam_object_detect(self):
         self.image_path = ""
         result_box = self.root.ids.cam_detect_box.ids.cam_result_image
+        result_box.clear_widgets()
+
+    def reset_classify(self):
+        self.image_path = ""
+        uploaded_image_box = self.root.ids.img_classify_box.ids.uploaded_image
+        uploaded_image_box.clear_widgets()
+        result_box = self.root.ids.img_classify_box.ids.result_label
         result_box.clear_widgets()
 
     def open_link(self, instance, url):