+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Changes per version
 
+### 0.0.5
+- Adding `Image Prediction` menu which can give top 5 predictions out of 1000 types as per `Synset` list for the main object on image.
+- Changing the model files downlaod style to ask for user prompt.
+- Some minor improvements to load the app faster.
+
 ### 0.0.4
 - Adding a new menu to capture image from camera & detect objects (can have bugs)
 - Fix back button action in downlaod manager